refactor(lml): move conversions/types into lib/features/lml, fix lib→src inversion
lib/features/lml/api.ts imported its conversion function and response type
from src/hooks/lml/ — a state-layer module depending on the presentation
layer. Move lml-conversions.ts and types.ts (+ their colocated tests) into
lib/features/lml/, and useLmlLibrarySearch.ts (+ test) up to src/hooks/ as
a plain hook. Update the four importers (flowsheetHooks.ts + its test,
useLmlLibrarySearch.test.ts, lib/test-utils/fixtures.ts) and drop the
src/hooks/lml/index.ts barrel, which had no single equivalent location
once its three re-exports landed in different directories.

Moves are verbatim (git-detected renames); only import paths changed.
Debounce/skip semantics (#625, #563) untouched. No lib→src imports remain
under lib/features/lml/; no src/hooks/lml/ directory remains. Test count
unchanged: 269 files / 3,670 tests.
